update types on functions
